Made Music V2 the new version
