+++ functions/music.js
@@ -1,52 +1,37 @@
 const ytdl = require('ytdl-core');
 const { youtube } = require('scrape-youtube');
 const { MessageEmbed } = require('discord.js');
-const skip = require('./music functions/skip');
 const stop = require('./music functions/stop');
+const skip = require('./music functions/skip');
+const queueFunc = require('./music functions/queue');
 const previous = require('./music functions/previous');
-const queue = require('./music functions/queue');
 const remove = require('./music functions/remove');
 
-// Global Vars
-let songObj = {
-    songPlaying: false,
-    currentSong: null,
-    songArray: [],
-    previousSongs: [],
-    prevSong: null,
-    prevCalled: false
-};
-
-let connection;
+const serverMap = new Map();
 
 module.exports = {
-    // name: ['play', 'p', 'skip', 'stop', 'leave', 'previous', 'prev', 'queue', 'remove'],
-    description: 'play song in discord channel',
+    name: ['play', 'p', 'stop', 'leave', 'skip', 'queue', 'q', 'previous', 'prev', 'remove'],
+    description: 'music thing',
     async execute(message, args) {
-        // Local Vars
         const vc = message.member.voice.channel;
         const command = message.content.substring(1).trim().split(/ +/).shift().toLowerCase();
-        let incomingVideo;
 
         if (!vc) return message.reply('You gotta be in a voice channel');
 
         const perms = vc.permissionsFor(message.client.user);
-
+        
         if (!perms.has('CONNECT')) return message.reply('You\'re missing the CONNECT permission');
 
-        if (!perms.has('SPEAK')) return message.reply('You\'re missing the SPEAK permission fool');       
+        if (!perms.has('SPEAK')) return message.reply('You\'re missing the SPEAK permission fool');
 
-        // Everything for the play command
         if (command == 'play' || command == 'p') {
             if (!args.length) return message.reply('You need to add another argument bro');
 
-            let testArg = ytdl.validateURL(args.join(' '));
+            let queue = serverMap.get(message.guild.id);
 
-            if (testArg) {
-                incomingVideo = ytdl.getURLVideoID(args.join(' '));
-            } else {
-                incomingVideo = `${args.join(' ')} audio`;
-            }
+            let checkArg = ytdl.validateURL(args[0]);
+
+            let video;
 
             const findVideo = async (query) => {
                 const result = await youtube.search(query);
@@ -54,53 +39,44 @@ module.exports = {
                 return (result.videos.length >= 1) ? result.videos[0] : null;
             }
 
-            const video = await findVideo(incomingVideo);
+            const playSong = async(guildId, song) => {
+                const server = serverMap.get(guildId);
 
-            if (!songObj.songPlaying) {
-                connection = await vc.join();
-                songObj.songPlaying = true;
-            } else {
-                songObj.songArray.push(video);
+                if (!server.stopCalled) {
+                    server.currentSong = song;
+                }
 
-                const queueEmbed = new MessageEmbed()
-                    .setColor([2, 150, 255])
-                    .setAuthor(`Added ${video.title} to the queue 👍`)
-                    .setThumbnail(video.thumbnail)
-                    .setTimestamp()
-                    .setFooter(`${message.author.username}`, message.author.avatarURL({ dynamic:true }));
+                server.songArray.shift();
 
-                message.channel.send(queueEmbed);
-                return;
-            }
-
-            const playSong = async(song) => {
-                songObj.currentSong = song
-
-                const stream = ytdl(song.link, {filter: 'audioonly', quality: 'lowestaudio', highWaterMark: 1<<25}).on('error', err => {
-                    console.log(err);
-                    vc.leave();
-                });
-                connection.play(stream, {seek: 0, volume: 1})
-                    .on('finish', () =>{
-                        if(songObj.songArray.length == 0) {
-                            songObj.songPlaying = false;
-                            songObj.previousSongs = [];
-                            songObj.songArray = [];
-                            vc.leave();
-                        } else {
-                            songObj.prevSong = songObj.currentSong;
-                            songObj.previousSongs.unshift(songObj.prevSong);
-                            playSong(songObj.songArray[0]);
-                            songObj.songArray.shift()
-
-                            if (songObj.prevCalled) {
-                                songObj.previousSongs.shift()
-                                songObj.prevCalled = false;
-                            }
-                        }
+                if (!song) {
+                    server.voice.leave();
+                    serverMap.delete(guildId);
+                    return;
+                } else {
+                    const stream = ytdl(song.link, {filter: 'audioonly', quality: 'lowestaudio', highWaterMark: 1<<25}).on('error', err => {
+                        console.log(err);
+                        server.text.send('There was an error with that stream');
+                        server.voice.leave();
                     });
-        
-                    const nowPlaying = new MessageEmbed()
+                    server.connection.play(stream, {seek: 0, volume: 1})
+                        .on('finish', () => {
+                            if(server.songArray.length == 0) {
+                                server.voice.leave();
+                                serverMap.delete(guildId);
+                            } else {
+                                server.prevSong = server.currentSong;
+                                server.previousSongs.unshift(server.prevSong);
+                                playSong(guildId, server.songArray[0]);
+
+                                if (server.prevCalled) {
+                                    server.previousSongs.shift()
+                                    server.prevCalled = false;
+                                }
+                            }
+                        })
+                }
+
+                const nowPlaying = new MessageEmbed()
                         .setColor([255, 0, 255])
                         .setAuthor('Tilly Music Player')
                         .addFields(
@@ -110,48 +86,117 @@ module.exports = {
                         .setThumbnail(song.thumbnail)
                         .setTimestamp()
             
-                await message.channel.send(nowPlaying);
+                await server.text.send(nowPlaying);
             }
 
-            if (video) {
-                playSong(video);
+            if (checkArg) {
+                let info = await ytdl.getBasicInfo(args[0]);
+
+                video = {
+                    title: info.videoDetails.title,
+                    thumbnail: info.videoDetails.thumbnails[0].url,
+                    link: info.videoDetails.video_url
+                }
             } else {
-                message.reply(`I couldn't find a video for that fam`);
+                let foundVideo = await findVideo(`${args.join(' ')} audio`);
+
+                video = {
+                    title: foundVideo.title,
+                    thumbnail: foundVideo.thumbnail,
+                    link: foundVideo.link
+                }
+            }
+
+            if (!queue) {
+                const songObj = {
+                    voice: vc,
+                    connection: null,
+                    text: message.channel,
+                    currentSong: null,
+                    songArray: [],
+                    previousSongs: [],
+                    prevSong: null,
+                    prevCalled: false,
+                    stopCalled: false
+                };
+
+                serverMap.set(message.guild.id, songObj);
+                songObj.songArray.push(video);
+                
+                try {
+                    songObj.connection = await vc.join();
+                    playSong(message.guild.id, songObj.songArray[0]);
+                } catch (error) {
+                    serverMap.delete(message.guild.id);
+                    message.channel.send('Had trouble joining the voice channel');
+                    throw error;
+                }
+
+            } else {
+                queue.songArray.push(video);
+
+                const queueEmbed = new MessageEmbed()
+                    .setColor([2, 150, 255])
+                    .setAuthor(`Added ${video.title} to the queue 👍`)
+                    .setThumbnail(video.thumbnail)
+                    .setTimestamp()
+                    .setFooter(`${message.author.username}`, message.author.avatarURL({ dynamic:true }));
+
+                queue.text.send(queueEmbed);
                 return;
             }
         }
 
-
-        // Everything for the skip command
-        if (command == 'skip') {
-            if (!vc) return message.reply('You gotta be in a voice channel bro');
-
-            skip(songObj, connection, message);
-        }
-
-
-        // Everything for the stop / leave command
         if (command == 'stop' || command == 'leave') {
-            if (!vc) return message.reply('You gotta be in a voice channel');
-            
-            stop(songObj, connection, message);
+            let guild = serverMap.get(message.guild.id);
+
+            if (!vc) return message.reply('You gotta be in the voice channel fam');
+
+            if (!guild) return message.reply('I\'m not even playing anything');
+
+            guild.stopCalled = true;
+
+            stop(guild);
+
+            serverMap.delete(message.guild.id);
         }
 
-        // Everything for the previous song command
-        if (command == 'previous' || command == 'prev') {
-            if (!vc) return message.reply('You gotta be in a voice channel');
+        if (command == 'skip') {
+            let guild = serverMap.get(message.guild.id);
 
-            previous(songObj, connection, message);
+            if (!vc) return message.reply('You gotta be in the voice channel fam');
+
+            if (!guild) return message.reply('I\'m not even playing anything');
+
+            skip(serverMap.get(message.guild.id));
+        }
+
+        if (command == 'previous' || command == 'prev') {
+            let guild = serverMap.get(message.guild.id);
+
+            if (!vc) return message.reply('You gotta be in the voice channel fam');
+
+            if (!guild) return message.reply('I\'m not even playing anything');
+
+            previous(guild);
         }
 
         if (command == 'queue' || command == 'q') {
-            if (!vc) return message.reply('You gotta be in a voice channel');
+            let guild = serverMap.get(message.guild.id);
 
-            queue(songObj, message);
+            if (!vc) return message.reply('You gotta be in the voice channel fam');
+
+            if (!guild) return message.reply('Nothing\'s playing right meow');
+
+            queueFunc(serverMap.get(message.guild.id));
         }
 
         if (command == 'remove') {
-            if (!songObj.songPlaying) return message.reply('I\'m not playing anything right meow');
+            let guild = serverMap.get(message.guild.id);
+
+            if (!vc) return message.reply('You gotta be in the voice channel fam');
+
+            if (!guild) return message.reply('Nothing\'s playing right meow');
 
             if (args.length == 0) return message.reply('You didn\'t send enough arguments for me to work with. Try also sending the number of the song you want me to remove.');
 
@@ -159,9 +204,9 @@ module.exports = {
 
             if (args.length > 1) return message.reply('I can only remove 1 song at the moment');
 
-            if (songObj.songArray.length <= 0) return message.reply('There\'s nothing in the queue for me to remove');
+            if (guild.songArray.length <= 0) return message.reply('There\'s nothing in the queue for me to remove');
 
-            remove(songObj, message, parseInt(args[0]));
+            remove(guild, parseInt(args[0]));
         }
     }
 }
